+++ R/dev-example.r
@@ -32,9 +32,19 @@ dev_example <- function(topic, quiet = FALSE) {
 #' @param macros Custom macros to use to parse the `.Rd` file. See the
 #'   `macros` argument of [tools::parse_Rd()]. If `NULL`, then the
 #'   [tools::Rd2ex()] (and [tools::parse_Rd()]) default is used.
+#' @param run,test Deprecated, see `run_dontrun` and `run_donttest` above.
 run_example <- function(path, run_donttest = FALSE, run_dontrun = FALSE,
                         env = new.env(parent = globalenv()),
-                        quiet = FALSE, macros = NULL) {
+                        quiet = FALSE, macros = NULL, run, test) {
+
+  if (!missing(run)) {
+    warning("`run_example(run=)` is deprecated, please use `run_example(run_dontrun=)` instead", call. = FALSE)
+    run_dontrun <- run
+  }
+  if (!missing(test)) {
+    warning("`run_example(test=)` is deprecated, please use `run_example(run_donttest=)` instead", call. = FALSE)
+    run_donttest <- test
+  }
 
   if (!file.exists(path)) {
     stop("'", path, "' does not exist", call. = FALSE)